intellistack/backend/requirements.txt
# Core framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Authentication & Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
authlib>=1.3.0
httpx>=0.26.0

# AI/ML
openai>=1.10.0
langchain>=0.1.0
langchain-openai>=0.0.5
qdrant-client>=1.7.0
cohere>=4.45

# Utilities
structlog>=24.1.0
redis>=5.0.0
uuid-utils>=0.6.0
python-dotenv>=1.0.0
email-validator>=2.1.0

# Validation
phonenumbers>=8.13.0
